microsoft-agent-framework-sample/requirements.txt
agent-framework==1.0.0
python-dotenv
azure-ai-documentintelligence==1.0.2
openai
azure-core
azure-identity  
streamlit